reorganize shims by platform
